multiplayer-game/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.25
asyncpg==0.29.0
python-jose[cryptography]==4.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
websockets==12.0
pydantic-settings==2.2.1
